requirements.txt
streamlit
google-generativeai
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
os
requests
beautifulsoup4